% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello.name}
\alias{hello.name}
\title{Say hello to anybody}
\usage{
hello.name(name)
}
\arguments{
\item{name}{A character specifying who you want to greet.}
}
\value{
Character.
}
\description{
This function returns the string "Hello, *name*!" where name is the input of
type character that the user needs to input.
}
\note{
An error is thrown if the input name is not a character.
}
\examples{
hello.name("Ann")
x <- hello.name("Christian")
x

}
