% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello to the world}
\usage{
hello()
}
\value{
Character "Hello, world!"
}
\description{
Function without any arguments which returns the string "Hello, world!".
}
\examples{
hello()

}
